cleaned up and renamed files
